# Project/swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "Swagger Petstore"
  contact:
    email: "danielmcoelho@ua.pt"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:5000"
basePath: "/"
...
schemes:
- "https"
- "http"
paths:
...
  /account/{id}/payment:
    post:
      tags:
      - "payment"
      summary: "Create a payment"
      operationId: "create_payment"
      produces:
      - "application/json"
      parameters:
      - name : "id"
        in : "path"
        required: true
        type: "string"
        format: "uuid"
        description: "The identifier number of the account"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Payment"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Pet not found"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Payment:
    type: "object"
    properties:
      id:
        type: "string"
        format: "uuid"
        description: "A payment identifier number"
      request_id:
        type: "string"
        format: "uuid"
        description: "A product identifier number"
      buyer_id:
        type: "string"
        format: "uuid"
        description: "The buyer account identifier"
      seller_id:
        type: "string"
        format: "uuid"
        description: "The seller account identifier"
      created_at:
        type: "string"
        format: "date-time"
        description: "When the payment process was created"
      completed_at:
        type: "string"
        format: "date-time"
        description: "When the payment process was updated"
      state:
        type: "string"
        description: "Payment Status"
        enum:
        - "pending"
        - "completed"
        - "cancelled"    
      amount:
        type: "number"
        format: "float"
        description: "The total amount to pay for the product"
      currency:
        type: "string"
        description: "international standard that defines three-letter codes as currencies established by the International Organization. (ISO 4217)"    
      reference:
        type: string
        description: "A brief description about the product"
    xml:
      name: "Payment"
...